+++ CHANGELOG.md
@@ -17,6 +17,7 @@
 - New #199: Add `media` field to `PollOption` type.
 - New #199: Add `InputPollMedia` and `InputPollOptionMedia` types.
 - New #199: Add `explanationMedia`, `media`, `membersOnly` and `countryCodes` parameters to `SendPoll` method.
+- New #199: Add `media` field to `InputPollOption` type.
 
 ## 0.20 April 26, 2026